services/vector-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
numpy==2.2.3
faiss-cpu==1.10.0
openai==1.66.3
pydantic>=2.0.0
requests==2.31.0
